Changed charLocation from short to int to fix overflow in font groups longer than 65535 pixels

+++ main.c
@@ -115,7 +115,7 @@ void main()
     unsigned short numOfChars[fontWidthNum];
     unsigned short numPerWidth[fontWidthNum];
 
-    unsigned short charLocation[fontWidthNum][count];
+    unsigned int charLocation[fontWidthNum][count];
     unsigned short charName[fontWidthNum][count];
     unsigned char charWeight[fontWidthNum][count];
     unsigned char charHeight[fontWidthNum][count];
